feat(refactor): :recycle: split lines module
