+++ src/client/fileSimulation.tsx
@@ -5,6 +5,7 @@ import { createPortal, flushSync } from 'react-dom';
 import { FileCircle, type FileCircleHandle } from './components/FileCircle';
 import { PhysicsProvider } from './hooks';
 import * as Physics from './physics';
+import { computeScale } from './scale';
 const { Body, Composite, Engine } = Physics;
 
 const MIN_CIRCLE_SIZE = 1;
@@ -12,95 +13,12 @@ const CHAR_ANIMATION_MS = 1500;
 export const EFFECT_DROP_THRESHOLD = 50;
 export const MAX_EFFECT_CHARS = 100;
 
-const fileColors: Record<string, string> = {
-  '.ts': '#2b7489',
-  '.js': '#f1e05a',
-  '.json': '#292929',
-  '.md': '#083fa1',
-  '.html': '#e34c26',
-  '.css': '#563d7c',
-};
-
-const hashHue = (s: string): number =>
-  Array.from(s).reduce((a, c) => a + c.charCodeAt(0), 0) % 360;
-
-const hexToHsl = (
-  hex: string,
-): { h: number; s: number; l: number } => {
-  const r = parseInt(hex.slice(1, 3), 16) / 255;
-  const g = parseInt(hex.slice(3, 5), 16) / 255;
-  const b = parseInt(hex.slice(5, 7), 16) / 255;
-  const max = Math.max(r, g, b);
-  const min = Math.min(r, g, b);
-  const l = (max + min) / 2;
-  if (max === min) return { h: 0, s: 0, l: l * 100 };
-  const d = max - min;
-  const s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
-  let h = 0;
-  switch (max) {
-    case r:
-      h = (g - b) / d + (g < b ? 6 : 0);
-      break;
-    case g:
-      h = (b - r) / d + 2;
-      break;
-    default:
-      h = (r - g) / d + 4;
-  }
-  h *= 60;
-  return { h, s: s * 100, l: l * 100 };
-};
-
-const hsl = ({ h, s, l }: { h: number; s: number; l: number }): string =>
-  `hsl(${h},${s}%,${l}%)`;
-
-export const colorForFile = (name: string): string => {
-  const i = name.lastIndexOf('.');
-  const ext = i >= 0 ? name.slice(i).toLowerCase() : '';
-  const offset = (hashHue(name) % 20) - 10;
-  const base = fileColors[ext];
-  if (base) {
-    const { h, s, l } = hexToHsl(base);
-    return hsl({ h: (h + offset + 360) % 360, s, l });
-  }
-  const hue = (hashHue(ext) + offset + 360) % 360;
-  return `hsl(${hue},60%,60%)`;
-};
-
-
 interface BodyInfo {
   el: HTMLElement;
   body?: Physics.Body;
   r: number;
   handle?: FileCircleHandle;
 }
-
-export const computeScale = (
-  width: number,
-  height: number,
-  data: LineCount[],
-  opts: { linear?: boolean } = {},
-): number => {
-  if (!Array.isArray(data)) return 0;
-  const exp = opts.linear ? 1 : 0.5;
-  const maxLines = data.reduce(
-    (m, d) => Math.max(m, Number.isFinite(d.lines) ? d.lines : 0),
-    1,
-  );
-  const base = Math.min(width, height) / Math.pow(maxLines, exp);
-  const totalArea = data.reduce(
-    (sum, f) => {
-      const lines = Number.isFinite(f.lines) ? f.lines : 0;
-      return sum + Math.PI * ((Math.pow(lines, exp) * base) / 2) ** 2;
-    },
-    0,
-  );
-  const ratio = totalArea / (width * height);
-  const threshold = 0.15;
-  if (!Number.isFinite(ratio) || ratio <= threshold) return base;
-  const easing = Math.pow(threshold / ratio, 0.25);
-  return base * easing;
-};
 
 export const createFileSimulation = (
   container: HTMLElement,
